+++ BankService/src/main/java/com/example/service/impl/MoneyAccountServiceImpl.java
@@ -1,6 +1,6 @@
 package com.example.service.impl;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 import com.example.entity.MoneyAccount;
 import com.example.repository.MoneyAccountRepository;
@@ -21,13 +21,13 @@ public class MoneyAccountServiceImpl implements MoneyAccountService {
         if (moneyAccountRepository.existsByUserId(userId)) {
             return;
         }
-        MoneyAccount moneyAccount = new MoneyAccount(userId, BigInteger.ZERO);
+        MoneyAccount moneyAccount = new MoneyAccount(userId, BigDecimal.ZERO);
         moneyAccountRepository.save(moneyAccount);
     }
 
     @Transactional
     @Override
-    public void addMoney(String userId, BigInteger amount) {
+    public void addMoney(String userId, BigDecimal amount) {
         MoneyAccount account = moneyAccountRepository.findByUserId(userId)
                 .orElseThrow(() -> new NoSuchEntityException("No account for user " + userId));
         account.setBalance(account.getBalance().add(amount));
