+++ BookService/src/main/java/com/example/dao/BookInfoDao.java
@@ -1,18 +1,18 @@
 package com.example.dao;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public class BookInfoDao {
     public String shortDescription;
     public String author;
     public String title;
     public boolean inModeration;
-    public BigInteger price;
+    public BigDecimal price;
 
     public BookInfoDao() {
     }
 
-    public BookInfoDao(String shortDescription, String author, String title, boolean inModeration, BigInteger price) {
+    public BookInfoDao(String shortDescription, String author, String title, boolean inModeration, BigDecimal price) {
         this.shortDescription = shortDescription;
         this.author = author;
         this.title = title;
