+++ SubscriptionService/src/main/java/com/example/component/impl/CountPriceCalculator.java
@@ -1,6 +1,6 @@
 package com.example.component.impl;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -15,17 +15,17 @@ public class CountPriceCalculator implements PriceCalculator {
             10, 10.0,
             30, 20.0
     );
-    private static final BigInteger HUNDRED = new BigInteger("100");
+    private static final BigDecimal HUNDRED = new BigDecimal("100");
 
     @Override
-    public BigInteger calculateSubscriptionPrice(List<BookInfoPriceDao> prices) {
-        BigInteger sum = prices.stream().map(dao -> dao.price).reduce(BigInteger.ZERO, BigInteger::add);
+    public BigDecimal calculateSubscriptionPrice(List<BookInfoPriceDao> prices) {
+        BigDecimal sum = prices.stream().map(dao -> dao.price).reduce(BigDecimal.ZERO, BigDecimal::add);
         double actualPercent = 100.0 - DISCOUNTS_THRESHOLD.keySet().stream()
                 .filter(cnt -> cnt <= prices.size())
                 .max(Integer::compareTo)
                 .map(DISCOUNTS_THRESHOLD::get)
                 .orElse(0.0);
-        BigInteger mul = new BigInteger(Double.toString(actualPercent)).divide(HUNDRED);
+        BigDecimal mul = new BigDecimal(Double.toString(actualPercent)).divide(HUNDRED);
         return sum.multiply(mul);
     }
 }
