+++ SubscriptionService/src/main/java/com/example/entity/UserSubscription.java
@@ -1,6 +1,6 @@
 package com.example.entity;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 
@@ -25,7 +25,7 @@ public class UserSubscription {
     @JdbcTypeCode(SqlTypes.JSON)
     @Column(name = "book_ids", columnDefinition = "jsonb")
     private List<String> bookIds;
-    private BigInteger price;
+    private BigDecimal price;
 
     public String getId() {
         return id;
@@ -67,11 +67,11 @@ public class UserSubscription {
         this.bookIds = bookIds;
     }
 
-    public BigInteger getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(BigInteger price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 }
