+++ SubscriptionService/src/main/java/com/example/service/SubscriptionService.java
@@ -1,9 +1,9 @@
 package com.example.service;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface SubscriptionService {
-    BigInteger getSubscriptionPriceByBooks(List<String> bookIds);
+    BigDecimal getSubscriptionPriceByBooks(List<String> bookIds);
     boolean createSubscription(String userId, List<String> bookIds, boolean infinite);
 }
